+++ apps/frontend/src/context/pageTitleContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+
+export type PageTitleContextValue = {
+  setPageTitle: (title: string | null) => void;
+};
+
+export const PageTitleContext = createContext<PageTitleContextValue | null>(null);
